Image errors fix

+++ controllers/project.controller.js
@@ -39,7 +39,7 @@ module.exports = {
         let previousProject = await ProjectsModel.findById(id)
         if (previousProject.image) {
             fs.unlink(`./images/${previousProject.image}`, err => {
-                if (err) next(err);
+                if (err) console.error(err);
             });
         }
         await ProjectsModel.findByIdAndRemove(id, (err, result) => {
@@ -58,7 +58,7 @@ module.exports = {
                 if (previousProjects.image) {
                     fs.unlink(`./images/${previousProjects.image}`, function (err) {
                         if (err) {
-                            return next(err)
+                            console.error(err);
                         }
                     });
                 }
@@ -96,9 +96,11 @@ module.exports = {
         let projects = req.body;
         if (projects.imageData) {
             let previousProject = await ProjectsModel.findById(id)
-            fs.unlink(`./images/${previousProject.image}`, err => {
-                if (err) return next(err);
-            });
+            if (previousProject.image) {
+                fs.unlink(`./images/${previousProject.image}`, err => {
+                    if (err) console.error(err);
+                });
+            }
             let data = projects.imageData.replace(/^data:image\/\w+;base64,/, "");
             let buf = new Buffer(data, 'base64');
             let timeStamp = (new Date()).getTime()
@@ -128,7 +130,9 @@ module.exports = {
     },
 
     async publishProject(req, res, next) {
-        await ProjectsModel.findByIdAndUpdate(req.params.id, {isPublic: req.body.isPublic})
+        await ProjectsModel.findByIdAndUpdate(req.params.id, {
+                isPublic: req.body.isPublic
+            })
             .then(() => ProjectsModel.findById(req.params.id))
             .then((result) => {
                 res.status(200).json({
